requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
Django==1.10.4
django-annoying==0.10.3
django-celery-beat==1.0.1
kombu==4.0.2
MySQL-python==1.2.5
oauthlib==2.0.1
python-dateutil==2.6.0
pytz==2016.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
tweepy==3.5.0
vine==1.1.3
